refactor: 구현 불가능한 부분 삭제 & Todo 컴포넌트 리팩토링
